+++ profiler_gui/common_functions.h
@@ -199,10 +199,10 @@ QString timeStringIntNs(TimeUnits _units, ::profiler::timestamp_t _interval);
 
 //////////////////////////////////////////////////////////////////////////
 
-QString shortenCountString(size_t count, int precision = 1);
+QString shortenCountString(uint64_t count, int precision = 1);
 QString shortenCountString(uint32_t count, int precision = 1);
 QString shortenCountString(int64_t count, int precision = 1);
-QString shortenCountString(int count, int precision = 1);
+QString shortenCountString(int32_t count, int precision = 1);
 
 //////////////////////////////////////////////////////////////////////////
